refactor: isolate sidebar overlays
